chore: move strategy UI into milestonecard and open milestone when adding strategy to it
